requirements.txt
asgiref==3.8.1
autopep8==1.5.7
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dj-database-url==0.5.0
Django==5.1.5
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-recaptcha==4.0.0
django-two-factor-auth==1.17.0
django_csp==3.8
et_xmlfile==2.0.0
gunicorn==20.1.0
h11==0.14.0
idna==3.10
numpy==2.2.2
openpyxl==3.1.5
pandas==2.2.3
phonenumbers==8.13.50
pillow==11.0.0
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.4
pytz==2021.1
qrcode==7.4.2
requests==2.32.3
six==1.17.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
Unipath==1.1
urllib3==2.2.3
uvicorn==0.31.1
whitenoise==5.3.0
